+++ src/dokumentgenerator/domainInterfaces.ts
